+++ app/controllers/projects_controller.rb
@@ -21,7 +21,6 @@ class ProjectsController < ApplicationController
     @project = Project.new(project_params)
     authorize @project
     assign_team_leader
-    @project.trust_ukprn = 10061021
 
     if @project.valid?
       @project.save
@@ -53,7 +52,7 @@ class ProjectsController < ApplicationController
   end
 
   private def project_params
-    params.require(:project).permit(:urn, :delivery_officer_id)
+    params.require(:project).permit(:urn, :trust_ukprn, :delivery_officer_id)
   end
 
   private def assign_team_leader