+++ spec/factories/academies_api/conversion_project.rb
@@ -1,6 +1,0 @@
-FactoryBot.define do
-  factory :academies_api_conversion_project, class: "AcademiesApi::ConversionProject" do
-    name_of_trust { "THE RETICULATED SPLINES TRUST" }
-    proposed_opening_date { "2022-04-01T00:00:00" }
-  end
-end