+++ spec/features/users_can_view_project_information_spec.rb
@@ -6,7 +6,7 @@ RSpec.feature "Users can view project information" do
   let(:project_id) { project.id }
 
   before do
-    mock_successful_api_responses(urn: 12345)
+    mock_successful_api_responses(urn: 12345, ukprn: 10061021)
     sign_in_with_user(user)
   end